pyproject.toml
[tool.poetry]
name = "todolist-messenger"
version = "0.1.0"
description = "Todo-list for messengers. Application for portfolio."
authors = ["Semyon Emelyanov <semseriou@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = {extras = ["speedups"], version = "^3.6.2"}
pymorphy2 = "^0.8"
pytest = "^5.4.1"
python-dotenv = "^0.12.0"
uvloop = "^0.14.0"
alembic = "^1.4.2"
sqlalchemy = "^1.3.16"
psycopg2-binary = "^2.8.5"
aiogram = "^2.7"
transliterate = "^1.10.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
